ensure all the structs include proper yaml labels
note that we need also to add the 'mapstructure' tag to the CACert
field to allow proper yaml decode from viper: viper takes care of
merging the configuration from the yaml file and uses mapstructure
for decoding the yaml file.
The only fields merged by viper are the URL and the CACert.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>