Drop setting a custom providerID
Not usre why we were using the elemental:// providerID while this should
be something like k3s://

Not only that but it was blocking the correct health check of the
clusters due to not finding the provider.

Also drops the provider config override from rke2 as it was setting it
to an empty value. We should be able to restore if it proves to be
needed but doesnt seem to.

Signed-off-by: Itxaka <igarcia@suse.com>