Set a fixed name config for rke/k3s deployments
Looks like the name change once rebooted was causing more problems
that what it looked like.
On the first deployment, some of the pods would be launched with
the hostname as part of them, i.e.
cloud-controller-manager-m-qemu-standard-pc-q35-ich9-2009-1d566d99-d81a-49ae-b0bb-3
and that was ok, but the deployment would also change the node
name to be rancher-NUMBERS which only took place after a restart.

The deployment would then try to recover, and bring up missing pods
with that name, but would also bring up the old pods with the
old name, causing some pods to fight for resources or
directly crash constantly.

This patch adds an extra config to the k3s/rk2 configurations
to set the node-name for the deployment agents and pods to the
inventory.Name, so no matter how many restarts or name changes
the machine suffers, the agents will always use that node-name
for any config that it needs.

Signed-off-by: Itxaka <igarcia@suse.com>