operator/register: rework the registration protocol
This commit changes the communication protocol between the operator and
the elemental-register client in order to take full advantage of the websocket
connection: SMBIOS data and labels are now passed via the websocket channel.
The protocol is now extensible and could be easily extended to pass arbitrary
data or request different kind of services to the operator.

This commit breaks backwards compatibility with older operators and
elemental-register clients.

Fixes #5

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>