operator: ensure MachineRegistrations have required config
the only real mandatory config we need is the target device (which will
be later cause host installation failure if not present): check its
presence at resource creation/change.
Note that an empty spec:config causes also a panic on the registration
yaml retrieval via the HTTP GET to the registration URL: this because
the config structure was not allocated (see #202).

Fixes #202

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>