Store hardare data as labels for nodes during registration
Use the ghw lib to obtain hardware data about the node and use it to
have a standard set of hardware labels on the registered inventory
during registration. Otherwise its not possible to obtain this data
afterwards unless you ssh into the machine and its even more difficult
to use selectors to include hardware stuff.

Signed-off-by: Itxaka <igarcia@suse.com>