Derive TPM seed from system UUID
Try first to get the system UUID and derive the seed from there. It
still should be unique per system but allows us to reproduce it if
needed.

This should be a stepping stone to deprecate setting the random seed
manually and resort to always random-seed by default when using emulated
TPM

Signed-off-by: Itxaka <igarcia@suse.com>