Return registration errors to client
Introduces two new message-types (MsgError and MsgConfig).

MsgError is sent when an error is encountered during the registration
process.

MsgConfig is used to send the elemental configuration to the client,
before this was just a raw message with no type so we need to check in
the server if the client supports the message, otherwise fallback to the
raw message.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>