operator: expose build-image API (#315)
* operator: add multiple APIs support in the API server callback

- parse incoming request to identify the requested API
- move the registration management in a separate function
- add placeholder for the new build-image API

* operator: move generic API functions from register.go to server.go

We are introducing new APIs: let's keep in the register.go package
only those functions specific to the register API only.

This commit just moves some functions from register.go to server.go.
No changes in the code.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: move getMachineRegistration() to server.go

Move getMachineRegistration() to server.go for usage from all
APIs. Moreover, let it take directly the token as parameter.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: enforce API syntax during registration

We expext to receive a path of the form:
/elemental/{api}
enforce it (or return HTTP 404 - Not found).

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: introduce generic getValue() function

This is a generic function that could stay under server.go.
Move the specific function to retrieve the CACert under register.go
and leverage the newly introduced function.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: rename register.go to api_registration.go

just to make code easier to navigate

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: add api_buildimage.go for build-image API functions

move there also the placeholder function for the build-image API

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: add registration cache to the server

We need it to store ephemeral build image data, like the seed image
and the status of the actual build job.
We could extend it in the future to have a full cache of the
MachineRegistration that the registration server should deal with.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: build-image API scaffolding

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: add tests for build-image api

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: ensure user input from APIs is properly escaped

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: log failure to set read deadline on the websocket

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* update Copyright year in modified files

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: add a small delay before reporting the build job failure

The build job is empty, and right now just reports failure (actual
implementation will be added in the future).
Since the API tests check the state just after starting the build and
expect to find its state updated to "Started", we need the build job to
wait a while before updating the build state to "Failed", otherwise the
tests may miss the Started state.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

* operator: make code scanning tools happier

when user input is sanitized, use a different var to store the sanitized
value. This should made scanning tools job eisier and avoid false
positives.
On the bonus side, the code will be more readable, i.e., it will be
clear where we use the sanitized values.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>