operator: register the host IP in MachineInventory annotations
This commit adds a new exchange between the registering client and the
operator: the registering client will pass some data that will be put in
the MachineInventory annotations.
This is meant to be a way to track those dynamic data from the host that
could be handy to have in the MachineInventory.
The only data passed in the current commit is the host address used to
register.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>