operator: build-image: retry build job Pod creation if needed
It may easily happen that when we try to create a new Pod for a build
process, there is an old job Pod which has the same name which is in the
Deletion phase: it is not over yet.
So, when creation fails, let's check if it is because a Pod with the
same name is already there an it is deleting: if so, wait a bit and
retry.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>