Sets a validation process for Machine Inventory adoption
Machine inventory selectors are all competing to adopt
a machine inventory resource. In order to avoid a machine
inventory being adopted by more than one selector or viceversa
this commit adds a validation process and tracks its status
in a new adoption condition for the inventory resource.

The validation process is simple: on one side the machine
inventory checks its owner includes the appropriate reference;
on the other side the selector checks the reference in status
points to a machine inventory owned by itself.

If an inconsistency is found during the validation the adoption
process restarts. The bootstrap plan is not created until the
adoption process is validated.

Signed-off-by: David Cassany <dcassany@suse.com>