SeedImage: inject MachineRegistration and date in the built iso name
Previously, the built iso name was a static 'elemental.iso'.
Now, the referenced MachineRegistration and the time in RFC3339 format
are part of the built iso name.

elemental-{MachineRegistration Name}-{RFC3339 date}.iso

es: elemental-myreg-2023-04-19T13:48:50.52Z.iso

Moreover, take advantage of the new httpfy functionality and pass it the
full name of the generated iso file to be served independently of the
URL in the request.

Drop also the static "elemental.iso" from the generated download URL.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>