goreleaser: fix releases CI
the SBOM files produced by goreleaser have the following default name
for binaries artifacts:
["{{ .Binary }}_{{ .Version }}_{{ .Os }}_{{ .Arch }}.sbom"]

This causes the SBOM of arm6 and arm7 to have the same name: this is
reported at the time to push the artifacts.

The result was that we just got one SBOM for the arm architecture, that
could be either related to arm6 or arm7, while the other one is dropped.

Just name the SBOM with the artifact name.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>