SeedImage: manage updates of builder Pod under deletion
If the Pod get scheduled to be deleted but the status is not updated
before the Pod reconcile loop, we may end up updating the status to
SeedImageReady to true as we found the Pod still running.
This commit adds a check if the Pod is marked for deletion: if so, let's
wait.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>