Charts: add a new chart to host the pre-hook migration template
We want to drop the pre-hook job reworking the annotations on the
Elemental Operator CRDs as soon as possible. The expected user workflow
is to upgrade to 1.2.4 first if they want to upgrade from the legacy
elemental-operator chart to the newer splitted charts.
So, let's drop the pre-hook job from the main elemental-operator chart
but let's keep it around, in a separate chart, that could still be used
as a standalone chart for manual migration.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>