Fix channel sync bug
Commit 8393423c8a0e93fc7865ef53879e35287c17a7db changes the
display-command to exit immediately, instead of hanging around.

When upgrading from an old version of elemental-operator to a newer, the
older version seems to still hang around and is used to spawn the
"pause"-container of the managed-os-version channel. This means
upgrading from 1.3.5 or earlier keeps the display-command that sleeps
forever which prevents the channel sync from completing.

This commit changes the channels to use 'cat' command instead of
'elemental-operator display --file ...'.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>