charts: backport changes from Rancher Marketplace chart
Mainly, change the var "channel.repository" to "channel.image":
"repository" var values are collected by the Rancher Marketplace for air
gap support. Anyway, in the case of the Elemental Channel, the image
should be re-generated in case of air gap scenario following the
official Elemental docs.

Also note that now the channel.image should contain the full url, not
prepended with teh registryUrl or the systemDefaultRegistry.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>