Allow cross build of seed-image ISOs
This commit makes use of the targetPlatform field on the seedImage spec
to allow building ISOs for different architectures.

It does this by spawning the initContainer using the seedImageBuilder
image and using the `elemental pull-image --platform=` command to
download the correct ISO and copies it to the attached volume.

One drawback of this approach is we don't get the caching of images in
the container runtime that we get when building natively.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>