register: don't send new Disks and Controllers data
The newer versions of jaypipes/ghw include a new type of Disk and
Controller in the Blocks section.

We use the library json serialization functionality: the deserialization
function of the older version of the library (0.9.0) would error out
when trying to decode serialized data of newer versions if they include
the new "virtual" drive or the new "loop" controller.
For now, just remove those devices.
Proper fix will be to better deal with this kind of errors avoiding
tearing down the registration process.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>